// frontend/src/core/types.ts
export interface User {
    id: number;
    created_at: string;
    last_seen: string;
    online: boolean;
    username: string;
    admin?: boolean;
    bio?: string;
}
// ...
export interface SendDMRequest {
    recipientId: number;
    iv: string;
    ciphertext: string;
    salt: string;
    iv2: string;
    wrappedMk: string;
}

// Responses

/**
 * Login response structure
 * @interface LoginResponse
 * @property {User} user - User information
 * @property {string} token - JWT authentication token
 */
export interface LoginResponse {
    user: User;
    token: string;
}

export interface BackupBlob {
    blob: string;
}

export interface FetchDMResponse {
    messages: DmEnvelope[]
}